refactor(infra): move SessionActorQueue to zeroclaw-infra
Move session_queue.rs from zeroclaw-gateway to zeroclaw-infra. The file
has zero gateway/Axum deps (pure tokio::sync + std::collections) and is
needed by the new runtime RPC layer (#6837).

Gateway's session_queue.rs becomes a re-export so all existing imports
continue to work.

Part of #6837.
