+++ crates/zeroclaw-gateway/src/session_queue.rs
@@ -1,234 +1,3 @@
-//! Per-session actor queue for serializing concurrent access.
-//!
-//! Each session gets at most one concurrent turn. Additional requests queue up
-//! (bounded by `max_queue_depth`) and proceed in FIFO order. This prevents
-//! SQLite history corruption from overlapping writes and ensures consistent
-//! session state transitions.
+//! Re-export from zeroclaw-infra so existing gateway imports keep working.
 
-use std::collections::HashMap;
-use std::sync::Arc;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::time::Duration;
-
-use tokio::sync::{Mutex, OwnedSemaphorePermit, Semaphore};
-use tokio::time::Instant;
-
-/// Per-session serialization queue.
-pub struct SessionActorQueue {
-    slots: Mutex<HashMap<String, Arc<SessionSlot>>>,
-    max_queue_depth: usize,
-    lock_timeout: Duration,
-    idle_ttl: Duration,
-}
-
-struct SessionSlot {
-    semaphore: Arc<Semaphore>,
-    last_active: Mutex<Instant>,
-    pending: AtomicUsize,
-}
-
-/// RAII guard that releases the session permit on drop.
-pub struct SessionGuard {
-    slot: Arc<SessionSlot>,
-    _permit: OwnedSemaphorePermit,
-}
-
-impl Drop for SessionGuard {
-    fn drop(&mut self) {
-        self.slot.pending.fetch_sub(1, Ordering::Relaxed);
-    }
-}
-
-/// Errors from the session queue.
-#[derive(Debug)]
-pub enum SessionQueueError {
-    /// Too many requests queued for this session.
-    QueueFull { session_id: String, depth: usize },
-    /// Timed out waiting for the session lock.
-    Timeout { session_id: String },
-}
-
-impl std::fmt::Display for SessionQueueError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::QueueFull { session_id, depth } => {
-                write!(
-                    f,
-                    "Session {session_id} queue full ({depth} pending requests)"
-                )
-            }
-            Self::Timeout { session_id } => {
-                write!(f, "Timed out waiting for session {session_id}")
-            }
-        }
-    }
-}
-
-impl std::error::Error for SessionQueueError {}
-
-impl SessionActorQueue {
-    /// Create a new queue with the given limits.
-    pub fn new(max_queue_depth: usize, lock_timeout_secs: u64, idle_ttl_secs: u64) -> Self {
-        Self {
-            slots: Mutex::new(HashMap::new()),
-            max_queue_depth,
-            lock_timeout: Duration::from_secs(lock_timeout_secs),
-            idle_ttl: Duration::from_secs(idle_ttl_secs),
-        }
-    }
-
-    /// Acquire exclusive access to a session. Blocks until the session is free
-    /// or the timeout expires. Returns a guard that releases on drop.
-    pub async fn acquire(&self, session_id: &str) -> Result<SessionGuard, SessionQueueError> {
-        let slot = {
-            let mut slots = self.slots.lock().await;
-            slots
-                .entry(session_id.to_string())
-                .or_insert_with(|| {
-                    Arc::new(SessionSlot {
-                        semaphore: Arc::new(Semaphore::new(1)),
-                        last_active: Mutex::new(Instant::now()),
-                        pending: AtomicUsize::new(0),
-                    })
-                })
-                .clone()
-        };
-
-        // Check queue depth before waiting
-        let current = slot.pending.fetch_add(1, Ordering::Relaxed);
-        if current >= self.max_queue_depth {
-            slot.pending.fetch_sub(1, Ordering::Relaxed);
-            return Err(SessionQueueError::QueueFull {
-                session_id: session_id.to_string(),
-                depth: current,
-            });
-        }
-
-        // Acquire owned permit with timeout
-        let sem = slot.semaphore.clone();
-        match tokio::time::timeout(self.lock_timeout, sem.acquire_owned()).await {
-            Ok(Ok(permit)) => {
-                *slot.last_active.lock().await = Instant::now();
-                Ok(SessionGuard {
-                    slot,
-                    _permit: permit,
-                })
-            }
-            Ok(Err(_)) | Err(_) => {
-                slot.pending.fetch_sub(1, Ordering::Relaxed);
-                Err(SessionQueueError::Timeout {
-                    session_id: session_id.to_string(),
-                })
-            }
-        }
-    }
-
-    /// Get the number of pending requests for a session.
-    pub async fn queue_depth(&self, session_id: &str) -> usize {
-        let slots = self.slots.lock().await;
-        slots
-            .get(session_id)
-            .map(|s| s.pending.load(Ordering::Relaxed))
-            .unwrap_or(0)
-    }
-
-    /// Remove idle session slots that haven't been accessed within the TTL.
-    pub async fn evict_idle(&self) -> usize {
-        let mut slots = self.slots.lock().await;
-        let now = Instant::now();
-        let before = slots.len();
-        let ttl = self.idle_ttl;
-
-        let mut to_remove = Vec::new();
-        for (key, slot) in slots.iter() {
-            let last = *slot.last_active.lock().await;
-            if now.duration_since(last) > ttl && slot.pending.load(Ordering::Relaxed) == 0 {
-                to_remove.push(key.clone());
-            }
-        }
-        for key in &to_remove {
-            slots.remove(key);
-        }
-
-        before - slots.len()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn serializes_same_session() {
-        let queue = SessionActorQueue::new(8, 5, 600);
-
-        // Acquire and release, then re-acquire should work
-        let guard1 = queue.acquire("s1").await.unwrap();
-        drop(guard1);
-        let _guard2 = queue.acquire("s1").await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn parallel_different_sessions() {
-        let queue = SessionActorQueue::new(8, 5, 600);
-        let _guard1 = queue.acquire("s1").await.unwrap();
-        let _guard2 = queue.acquire("s2").await.unwrap();
-        // Both acquired simultaneously — different sessions don't block each other
-    }
-
-    #[tokio::test]
-    async fn queue_depth_limit() {
-        let queue = Arc::new(SessionActorQueue::new(2, 30, 600));
-
-        // Hold the session lock (pending=1)
-        let guard = queue.acquire("s1").await.unwrap();
-
-        // Queue one more (pending=2, will block waiting for permit)
-        let queue_clone = queue.clone();
-        let handle = tokio::spawn(async move { queue_clone.acquire("s1").await });
-
-        // Give the spawned task time to register
-        tokio::time::sleep(Duration::from_millis(50)).await;
-
-        // Third request should be rejected (pending=2 >= max=2)
-        let result = queue.acquire("s1").await;
-        assert!(matches!(result, Err(SessionQueueError::QueueFull { .. })));
-
-        drop(guard);
-        let _ = handle.await;
-    }
-
-    #[tokio::test]
-    async fn timeout_returns_error() {
-        let queue = SessionActorQueue::new(8, 1, 600);
-        let _guard = queue.acquire("s1").await.unwrap();
-
-        let start = Instant::now();
-        let result = queue.acquire("s1").await;
-        assert!(matches!(result, Err(SessionQueueError::Timeout { .. })));
-        assert!(start.elapsed() >= Duration::from_millis(900));
-    }
-
-    #[tokio::test]
-    async fn idle_eviction() {
-        let queue = SessionActorQueue::new(8, 5, 0); // 0s TTL
-        {
-            let _guard = queue.acquire("s1").await.unwrap();
-        }
-        tokio::time::sleep(Duration::from_millis(10)).await;
-        let evicted = queue.evict_idle().await;
-        assert_eq!(evicted, 1);
-    }
-
-    #[tokio::test]
-    async fn queue_depth_reports_correctly() {
-        let queue = SessionActorQueue::new(8, 30, 600);
-        assert_eq!(queue.queue_depth("s1").await, 0);
-
-        let guard = queue.acquire("s1").await.unwrap();
-        assert_eq!(queue.queue_depth("s1").await, 1);
-
-        drop(guard);
-        assert_eq!(queue.queue_depth("s1").await, 0);
-    }
-}
+pub use zeroclaw_infra::session_queue::*;
