+++ crates/zeroclaw-infra/src/lib.rs
@@ -5,6 +5,7 @@
 pub mod acp_session_store;
 pub mod debounce;
 pub mod session_backend;
+pub mod session_queue;
 pub mod session_sqlite;
 pub mod session_store;
 pub mod stall_watchdog;
